chore: Update CSS and HTML files
