--- Tools/Data.js
@@ -1,39 +1,0 @@
-import { GrFormNextLink } from "react-icons/gr";
-
-// This data will be called in the main file to render the elements
-
-const elementObject = [
-    {
-        name: "fennec <br/>fox",
-        country: "India",
-        image: "images/fennec-fox.jpg",
-        more: "Know more",
-        icons: <GrFormNextLink />
-    },
-
-    {
-        name: "humpback <br/>whale",
-        country: "South Africa",
-        image: "images/fennec-fox.jpg",
-        more: "Know more",
-        icons: <GrFormNextLink />
-    },
-
-    {
-        name: "common brown <br/>baboon",
-        country: "South Africa",
-        image: "images/fennec-fox.jpg",
-        more: "Know more",
-        icons: <GrFormNextLink />
-    },
-
-    {
-        name: "spotted <br/>deer",
-        country: "Amazon",
-        image: "images/fennec-fox.jpg",
-        more: "Know more",
-        icons: <GrFormNextLink />
-    }
-]
-
-export default elementObject;
