+++ Tools/data.js
@@ -1,0 +1,39 @@
+import { GrFormNextLink } from "react-icons/gr";
+
+// This data will be called in the main file to render the elements
+
+const elementObject = [
+  {
+    name: "fennec <br/>fox",
+    country: "India",
+    image:
+      "https://images.pexels.com/photos/24252211/pexels-photo-24252211/free-photo-of-a-table-with-flowers-and-watercolor-paints.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1",
+    more: "Know more",
+  },
+
+  {
+    name: "humpback <br/>whale",
+    country: "South Africa",
+    image:
+      "https://images.pexels.com/photos/24252211/pexels-photo-24252211/free-photo-of-a-table-with-flowers-and-watercolor-paints.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1",
+    more: "Know more",
+  },
+
+  {
+    name: "common brown <br/>baboon",
+    country: "South Africa",
+    image:
+      "https://images.pexels.com/photos/24252211/pexels-photo-24252211/free-photo-of-a-table-with-flowers-and-watercolor-paints.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1",
+    more: "Know more",
+  },
+
+  {
+    name: "spotted <br/>deer",
+    country: "Amazon",
+    image:
+      "https://images.pexels.com/photos/24252211/pexels-photo-24252211/free-photo-of-a-table-with-flowers-and-watercolor-paints.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1",
+    more: "Know more",
+  },
+];
+
+export default elementObject;
