+++ index.html
@@ -9,10 +9,12 @@
 </head>
 <body>
     
+    <h1 class="heading">Photo Gallery</h1>
     <!-- Adding a JavaScript root, which will render all data-->
-    <div id="root"></div>
+    <div class="mainroot"></div>
 
     <!-- Adding a JavaScript  file -->
-    <script type="module" src="bundles.js"></script>
+    <script type="module" src="index.js"></script>
+    
 </body>
 </html>
